+++ src/config.ts
@@ -1,13 +1,9 @@
+/**
+ * Described in package.json
+ */
 export interface Config {
-  /**
-   * Jump behavior when cursor on empty line.
-   * Default is `nonempty`.
-   *
-   * - `nonempty`: Jump to next/previous non empty line.
-   * - `empty`: Jump to next/previous empty line
-   */
   emptyLineBehavior: "empty" | "nonempty";
-//   circle: boolean;
+  cyclic: boolean;
 }
 
 export interface ConfigItem<T> {