+++ src/extension.ts
@@ -2,28 +2,44 @@ import { commands, ExtensionContext, TextEditor, window } from "vscode";
 import { Config } from "./config";
 import { CursorNinja } from "./cursor_ninja";
 
-function handler<T>(
+async function handler<T>(
   context: ExtensionContext,
-  cb: (context: ExtensionContext, editor: TextEditor, config: Config) => T
-): T | undefined {
+  cb: (
+    context: ExtensionContext,
+    editor: TextEditor,
+    config: Config
+  ) => Promise<T>
+): Promise<T | undefined> {
   const editor = window.activeTextEditor;
   if (editor == null) {
     return;
   }
 
-  return cb(context, editor, { emptyLineBehavior: "nonempty" });
+  try {
+    return await cb(context, editor, {
+      emptyLineBehavior: "nonempty",
+      cyclic: true,
+    });
+  } catch (error) {
+    console.error(error);
+    throw error;
+  }
 }
 
 export function activate(context: ExtensionContext) {
   [
     commands.registerCommand("cursor-ninja.jumpIndentDown", () =>
-      handler(context, (...arg) => CursorNinja.from(...arg).jumpIndent("down"))
+      handler(context, async (...arg) =>
+        CursorNinja.from(...arg).jumpIndent("down")
+      )
     ),
     commands.registerCommand("cursor-ninja.jumpIndentUp", () =>
-      handler(context, (...arg) => CursorNinja.from(...arg).jumpIndent("up"))
+      handler(context, async (...arg) =>
+        CursorNinja.from(...arg).jumpIndent("up")
+      )
     ),
     commands.registerCommand("cursor-ninja.scrollToCenterCursor", () =>
-      handler(context, (...arg) =>
+      handler(context, async (...arg) =>
         CursorNinja.from(...arg).scrollToCenterCursor()
       )
     ),