+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.0.2
+
+- Cyclic jump when the cursor in the bottom/top in editor.
+- Improve empty line behavior.
+
 ## 0.0.1
 
-- Initial release
+- Initial release.
+- schema